Add strategy to read from datafiles
